+++ releasenotes.md
@@ -87,6 +87,9 @@
 - Overaligning structs while using `@packed` would cause incorrect lowering #3000
 - Splatting a literal into a typed vaarg, e.g. `test(...(int[2]){ 88, 99 }, a: 123)` could cause the compiler to crash.
 - `&some_global[0]` was incorrectly considered a global constant when `some_global` was a slice.
+- Taking the type of a macro identifier would give the wrong error. 
+- Taking the type of a `$$builtin` function would crash the compiler.
+- Wrong error message when trying to take the address of a `$$builtin` function.
  
 ## 0.7.10 Change list